First cut of workers running coroutines
Builds correct logic for coroutines to run on workers. Only tasks that
consume threadpool resources now increase the executing_count, and
instead they increase executing_coroutines. This is because otherwise it
won't decide to run more then n_threads coroutines (which somewhat
defeats the purpose).

Introduces a new metric to measure event-loop load, by timing the time
spent in select(). As a backup, the load is the number of currently
executing coroutine-tasks. This is mainly to have a sortable metric that
can be used for scheduling.

Problems:
1. Some supreme cheese to be able to see the concurrency, the 'thread'
assigned to this is a modulus of the hash. Realistically the current task
stream view isn't particularly suited to viewing coroutines, needs
some thought.
2. Currently in ensure_computing there is a check such that we won't move
any more tasks to executing if there is no threads available in the thread
pool. Currently this will also stop executing more coroutines if it's
full, so needs some further thought there.